Pre-render on a separate surface on desktop too
Rendering directly to window surface was quite slick for debug, but prevents the possibility to scale the window, which would clearly make the game more enjoyable on desktop as well. Now scaling is very easily possible by simply defining a different window size (it's going to be horribly stretched if the ratio 2:1 is not respected).
Soon: add scaling parameters and decent aspect ratio management.